requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
cffi==1.17.1
    # via soundfile
lxml==5.3.0
    # via python-pptx
natsort==8.4.0
    # via -r requirements.in
numpy==2.2.2
    # via soundfile
pillow==11.1.0
    # via python-pptx
pycparser==2.22
    # via cffi
pyee==12.1.1
    # via python-ffmpeg
python-ffmpeg==2.0.12
    # via -r requirements.in
python-pptx==1.0.2
    # via python-pptx-interface
python-pptx-interface==0.0.12
    # via -r requirements.in
soundfile==0.13.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pyee
    #   python-ffmpeg
    #   python-pptx
xlsxwriter==3.2.2
    # via python-pptx
